Reorganize hybrid execution functions into dedicated hybrid_execution.c and hybrid_execution.h files
