+++ src/hybrid/parse_hybrid.c
@@ -770,86 +770,3 @@ error:
 
   return NULL;
 }
-
-static int HREQ_BuildPipelineAndExecute(HybridRequest *hreq, RedisModuleCtx *ctx,
-                    RedisSearchCtx *sctx, QueryError *status) {
-  RedisSearchCtx *sctx1 = hreq->requests[0]->sctx;
-  RedisSearchCtx *sctx2 = hreq->requests[1]->sctx;
-
-  if (RunInThread()) {
-    // Multi-threaded execution path
-    StrongRef spec_ref = IndexSpec_GetStrongRefUnsafe(sctx->spec);
-
-    // Create a dummy AREQ for BlockQueryClient (it expects an AREQ but we'll use the first one)
-    AREQ *dummy_req = hreq->requests[0];
-    RedisModuleBlockedClient* blockedClient = BlockQueryClient(ctx, spec_ref, dummy_req, 0);
-
-    blockedClientHybridCtx *BCHCtx = blockedClientHybridCtx_New(hreq, blockedClient, spec_ref);
-
-    // Mark the requests as thread safe, so that the pipeline will be built in a thread safe manner
-    for (size_t i = 0; i < hreq->nrequests; i++) {
-      AREQ_AddRequestFlags(hreq->requests[i], QEXEC_F_RUN_IN_BACKGROUND);
-    }
-
-    const int rc = workersThreadPool_AddWork((redisearch_thpool_proc)HREQ_Execute_Callback, BCHCtx);
-    RS_ASSERT(rc == 0);
-
-    return REDISMODULE_OK;
-    } else {
-    // Single-threaded execution path
-
-    // Build the pipeline and execute
-    if (HybridRequest_BuildPipeline(hreq, hreq->hybridParams, status) != REDISMODULE_OK) {
-      return REDISMODULE_ERR;
-    } else {
-      HREQ_Execute(hreq, ctx, sctx);
-      return REDISMODULE_OK;
-    }
-  }
-}
-
-/**
- * Main command handler for FT.HYBRID command.
- *
- * Parses command arguments, builds hybrid request structure, constructs execution pipeline,
- * and prepares for hybrid search execution.
- */
-int hybridCommandHandler(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
-  // Index name is argv[1]
-  if (argc < 2) {
-    return RedisModule_WrongArity(ctx);
-  }
-
-  const char *indexname = RedisModule_StringPtrLen(argv[1], NULL);
-  RedisSearchCtx *sctx = NewSearchCtxC(ctx, indexname, true);
-  if (!sctx) {
-    QueryError status = {0};
-    QueryError_SetWithUserDataFmt(&status, QUERY_ENOINDEX, "No such index", " %s", indexname);
-    return QueryError_ReplyAndClear(ctx, &status);
-  }
-
-  StrongRef spec_ref = IndexSpec_GetStrongRefUnsafe(sctx->spec);
-  CurrentThread_SetIndexSpec(spec_ref);
-
-  QueryError status = {0};
-
-  HybridRequest *hybridRequest = parseHybridCommand(ctx, argv, argc, sctx, indexname, &status);
-  if (!hybridRequest) {
-    goto error;
-  }
-
-  if (HREQ_BuildPipelineAndExecute(hybridRequest, ctx, sctx, &status) != REDISMODULE_OK) {
-    goto error;
-  }
-
-  CurrentThread_ClearIndexSpec();
-  return REDISMODULE_OK;
-
-error:
-  if (hybridRequest) {
-    HybridRequest_Free(hybridRequest);
-  }
-
-  CurrentThread_ClearIndexSpec();
-  return QueryError_ReplyAndClear(ctx, &status);
-}
